refactor: merge iterators types into iters.rs
